# Pre-load fee schemes when loading offences